+++ packages/core/RNSentryAndroidTester/app/src/test/java/io/sentry/react/Utils.kt
@@ -3,6 +3,7 @@ package io.sentry.react
 import android.content.pm.PackageInfo
 import android.content.pm.PackageManager
 import com.facebook.react.bridge.ReactApplicationContext
+import com.facebook.react.bridge.ReactContext
 import org.mockito.ArgumentMatchers.anyInt
 import org.mockito.ArgumentMatchers.anyString
 import org.mockito.Mockito.mock
@@ -10,7 +11,7 @@ import org.mockito.kotlin.whenever
 
 class Utils {
     companion object {
-        fun createRNSentryModuleWithMockedContext(): RNSentryModuleImpl {
+        fun makeReactContextMock(): ReactApplicationContext {
             val packageManager = mock(PackageManager::class.java)
             val packageInfo = mock(PackageInfo::class.java)
 
@@ -18,9 +19,13 @@ class Utils {
             whenever(reactContext.packageManager).thenReturn(packageManager)
             whenever(packageManager.getPackageInfo(anyString(), anyInt())).thenReturn(packageInfo)
 
+            return reactContext
+        }
+
+        fun createRNSentryModuleWithMockedContext(): RNSentryModuleImpl {
             RNSentryModuleImpl.lastStartTimestampMs = -1
 
-            return RNSentryModuleImpl(reactContext)
+            return RNSentryModuleImpl(makeReactContextMock())
         }
     }
 }